minor log output changes to pass the type checking on OS X (where size_t is not an unsigned int)

+++ servlib/bundling/BundleDaemon.cc
@@ -233,7 +233,7 @@ BundleDaemon::handle_bundle_received(BundleReceivedEvent* event)
         fragmentmgr_->proactively_fragment(bundle, proactive_frag_threshold_);
     if (nfrags > 0) {
         log_debug("proactive fragmentation: %u byte bundle => %d fragments",
-                  payload_len, nfrags);
+                  (u_int)payload_len, nfrags);
         return;
     }
